ensure proper formatting of timestamps
Suricata uses a custom time format which requires a
separate formatting string for parsing. If this format
is not used during marshaling into JSON, a round-trip
is not possible, and a JSON once marshaled can not be
parsed again by surevego. Hence we need to also marshal
the timestamp value in the same fashion.